+++ CHANGELOG.md
@@ -7,11 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2020-07-21
+
+### Added
+
+- Service endpoint, queryies and query handlers for fetching multiple products by category ids
+
 ### Changed
 
 - Service endpoint for fetching multiple products by ids or handles pluralized
 - Queries and query handlers for fetching multiple products by ids or handles pluralized
-- Service endpoint, queryies and query handlers for fetching multiple products by category ids
 
 ## [0.1.2] - 2020-07-16
 
@@ -42,7 +47,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Health service using CQRS for status checks
 - In-memory backend providing default query handlers
 
-[unreleased]: https://github.com/SorenA/lightops-commerce-services-product/compare/0.1.2...develop
+[unreleased]: https://github.com/SorenA/lightops-commerce-services-product/compare/0.2.0...develop
+[0.2.0]: https://github.com/SorenA/lightops-commerce-services-product/tree/0.2.0
 [0.1.2]: https://github.com/SorenA/lightops-commerce-services-product/tree/0.1.2
 [0.1.1]: https://github.com/SorenA/lightops-commerce-services-product/tree/0.1.1
 [0.1.0]: https://github.com/SorenA/lightops-commerce-services-product/tree/0.1.0